iran_news_clustering_pipeline/requirements.txt
psycopg2-binary
python-dotenv
openai
schedule
fastapi
uvicorn
gunicorn
requests
certifi
